Finish fetch Data
