+++ client/src/pages/Staff/index.jsx
@@ -1,22 +1,13 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import Navbar from '../../components/Navbar';
 import { useSelector } from 'react-redux';
 import Sidebar from '../../components/Sidebar';
-import StaffHeader from '../../components/StaffHeader';
 import styled from 'styled-components';
-import StaffMember from '../../components/StaffMember';
-import userImage from '../../images/user.jfif';
-import axios from 'axios';
+import SingleStaff from '../../components/SingleStaff';
 
 const Staff = () => {
   const { open } = useSelector((state) => state.sidebar.value);
   const { userProjects } = useSelector((state) => state.userProjects.value);
-
-  useEffect(() => {
-    const getProjectMembers = async () => {
-      
-    }
-  })
 
   return (
     <div className='page-container'>
@@ -25,47 +16,9 @@ const Staff = () => {
         <Navbar title={`Staff`} />
         <section className='page-content'  style={{paddingBottom: 0}}>
           <Staffs className='staffs'>
-            <div className='single-staff'>
-              <StaffHeader name='Graduation Project' />
-              <div className='staff-memebers'>
-                <StaffMember
-                  user={{
-                    image: userImage,
-                    name: 'Dev.karam',
-                    role: 'Admin',
-                  }}
-                />
-                <StaffMember
-                  user={{
-                    image: userImage,
-                    name: 'Dev.karam',
-                    role: 'Admin',
-                  }}
-                />
-                <StaffMember
-                  user={{
-                    image: userImage,
-                    name: 'Dev.karam',
-                    role: 'Admin',
-                  }}
-                />
-                <StaffMember
-                  user={{
-                    image: userImage,
-                    name: 'Dev.karam',
-                    role: 'Admin',
-                  }}
-                />
-              </div>
-            </div>
             {
               userProjects.map(({id, name}) => (
-                <div className="single-staff" key={id}>
-                  <StaffHeader name={name} />
-                  <div className="staff-memebers">
-
-                  </div>
-                </div>
+                <SingleStaff key={id} project={{id, name}} />
               ))
             }
           </Staffs>
@@ -95,31 +48,6 @@ const Staffs = styled.div`
   &::-webkit-scrollbar-thumb {
     background-color: #53535e;
     border-radius: .4rem;
-  }
-
-  .single-staff {
-    min-width: 330px;
-    margin-right: 2rem;
-
-    .staff-memebers {
-      max-height: 66vh;
-      overflow: auto;
-      padding-right: 1rem;
-      padding-bottom: 0.5rem;
-
-      &::-webkit-scrollbar-track {
-        background-color: transparent;
-        border-radius: 1rem;
-      }
-      &::-webkit-scrollbar {
-        width: 5px;
-        background-color: transparent;
-      }
-      &::-webkit-scrollbar-thumb {
-        background-color: #323239;
-        border-radius: 1rem;
-      }
-    }
   }
 `;
